+++ project/src/pages/add-review-page.tsx
@@ -1,15 +1,13 @@
 import { Link, useParams } from 'react-router-dom';
-import { Fragment,  ChangeEvent, FormEvent, useState, useEffect } from 'react';
+import { useEffect } from 'react';
 
-import { useAppSelector, useAppDispatch } from '../hooks';
+import { useAppSelector } from '../hooks';
 
 import Logo from '../components/logo/logo';
-
-import { NewComment } from '../types/films';
+import Review from '../components/review/review';
 
 import { store } from '../store/index';
 import { fetchFilm } from '../store/api-actions';
-import {addCommentAction} from '../store/api-actions';
 
 function AddReviewPage(): JSX.Element {
   const {id:idFilm} = useParams();
@@ -18,27 +16,8 @@ function AddReviewPage(): JSX.Element {
     store.dispatch(fetchFilm(Number(idFilm)));
   }, [idFilm]);
 
-  const { id, name, backgroundImage, posterImage }  = useAppSelector((state) => state.filmCurrent);
-
-  const [comment, setComment] = useState('');
-  const [rating, setRating] = useState(0);
-
-  const dispatch = useAppDispatch();
-
-  const onSubmit = (commentData: NewComment) => {
-    dispatch(addCommentAction(commentData));
-  };
-
-  const handleSubmit = (evt: FormEvent<HTMLFormElement>) => {
-    evt.preventDefault();
-    if (comment) {
-      onSubmit({
-        filmId: id,
-        comment: comment,
-        rating: rating,
-      });
-    }
-  };
+  const { filmCurrent } = useAppSelector(({DATA}) => DATA);
+  const { id, name, backgroundImage, posterImage }  = filmCurrent;
 
   return (
     <section className="film-card film-card--full">
@@ -81,39 +60,8 @@ function AddReviewPage(): JSX.Element {
       </div>
 
       <div className="add-review">
-        <form action="#" className="add-review__form"
-          onSubmit={handleSubmit}
-        >
-          <div className="rating">
-            <div className="rating__stars">
-              {[...Array(10)].reverse().map((_,index) => {
-                const keyValue = 10-index;
-                return (
-                  <Fragment key={keyValue}>
-                    <input className="rating__input" id={`star-${keyValue}`} type="radio" name="rating" value={keyValue}
-                      defaultChecked={keyValue===rating}
-                      onChange={({target}: ChangeEvent<HTMLInputElement>) => { setRating(Number(target.value));}}
-                    />
-                    <label className="rating__label" htmlFor={`star-${keyValue}`}>{keyValue}</label>
-                  </Fragment>
-                );
-              })}
-            </div>
-          </div>
 
-          <div className="add-review__text">
-            <textarea className="add-review__textarea" name="comment" id="review-text" placeholder="Review text" onChange={(event) => setComment(event.target.value)} value={comment}></textarea>
-            <div className="add-review__submit">
-              <button
-                className="add-review__btn"
-                type="submit"
-                disabled={(!comment || rating===0)}
-              >Post
-              </button>
-            </div>
-
-          </div>
-        </form>
+        <Review id={id} />
       </div>
   );
     </section>
