+++ project/src/store/index.ts
@@ -1,5 +1,5 @@
 import {configureStore} from '@reduxjs/toolkit';
-import {reducer} from './reducer';
+import {rootReducer} from './root-reducer';
 import {createAPI} from '../services/api';
 
 import {redirect} from './middlewares/redirect';
@@ -7,7 +7,7 @@ import {redirect} from './middlewares/redirect';
 export const api = createAPI();
 
 export const store = configureStore({
-  reducer,
+  reducer: rootReducer,
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware({
       thunk: {
